#2 refactor: DIContainer 의존성 주입 코드를 AppCoordinator -> SeceneDelegate로 이동하여 응집도 개선
