+++ Projects/App/Sources/SceneDelegate.swift
@@ -7,6 +7,14 @@
 
 import UIKit
 
+import CoreDataMapper
+import CoreDIContainer
+import CoreNetworkService
+import CoreStorageService
+import FeatureLoginDIContainer
+import FeatureLoginRepository
+import FeatureLoginUseCase
+
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     var window: UIWindow?
@@ -20,11 +28,30 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
         
+        registerDependencies()
+        
         let rootNavigationController = UINavigationController()
         appCoordinator = DefaultAppCoordinator(navigationController: rootNavigationController)
         appCoordinator?.start()
         
         window?.rootViewController = rootNavigationController
         window?.makeKeyAndVisible()
+    }
+}
+
+// MARK: - Dependency Registration
+
+private extension SceneDelegate {
+    func registerDependencies() {
+        // MARK: - Core
+        CoreDIContainer.shared.register(service: AnyDataMapper.self) { AnyDataMapper(CoreExampleDataMapper()) }
+        CoreDIContainer.shared.register(service: EndpointService.self) { URLSessionEndpointService.shared }
+        CoreDIContainer.shared.register(service: URLDataService.self) { URLSessionURLDataService.shared }
+        CoreDIContainer.shared.register(service: DiskCacheStorage.self) { FileManagerDiskCacheStorage.shared }
+        CoreDIContainer.shared.register(service: MemoryCacheStorage.self) { NSCacheMemoryCacheStorage.shared }
+        
+        // MARK: - Login
+        LoginDIContainer.shared.register(service: LoginUseCase.self) { DefaultLoginUseCase() }
+        LoginDIContainer.shared.register(service: LoginRepository.self) { DefaultLoginRepository() }
     }
 }
